requirements.txt
slack-bolt==1.20.1
slack-sdk==3.33.4
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
